model/requirements.txt
flask

pandas
numpy

joblib
scikit-learn
sqlalchemy
mysql-connector-python
pymysql